Update templates for Textual 7
This is a pretty huge update that also adds compatibility for HSL
hashing for nickname colors. Enabling this in preferences makes all
instances of the user's own nickname uniform (inline and sender), and
allows the /setcolor command to be used to customize the style without
delving into CSS at all.